use uint32_t

+++ benchmarks/spectre/spectre-stl.c
@@ -8,8 +8,8 @@ extern uint32_t __VERIFIER_nondet_uint(void);
 #endif
 
 uint32_t publicarray_size = 16;
-uint8_t publicarray[16] = { 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15 };
-uint8_t publicarray2[16] = { 20 };
+uint32_t publicarray[16] = { 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15 };
+uint32_t publicarray2[16] = { 20 };
 
 // The attacker's goal in all of these examples is to learn any of the secret data in secretarray
 uint32_t secretarray_size = 16;
@@ -17,7 +17,7 @@ uint32_t secretarray[16] __attribute__((used)) = { 10,21,32,43,54,65,76,87,98,10
 
 
 // this is mostly used to prevent the compiler from optimizing out certain operations
-volatile uint8_t temp = 0;
+volatile uint32_t temp = 0;
 
 uint32_t idxg;
 
@@ -92,28 +92,28 @@ void victim_function_v4(uint32_t idx) { // INSECURE
 #endif
 
 #ifdef v5
-uint8_t *case5_ptr = secretarray;
+uint32_t *case5_ptr = secretarray;
 void victim_function_v5(uint32_t idx) {  // INSECURE
   register uint32_t ridx asm ("edx");
   ridx = idx % (publicarray_size - 1);
   
   case5_ptr = publicarray;       // Bypassed store
 
-  uint8_t toleak = case5_ptr[ridx];
+  uint32_t toleak = case5_ptr[ridx];
   temp &= publicarray2[toleak];
 }
 #endif
 
 #ifdef v6
 uint32_t case6_idx = 0;
-uint8_t *case6_array[2] = { secretarray, publicarray };
+uint32_t *case6_array[2] = { secretarray, publicarray };
 void victim_function_v6(uint32_t idx) { // INSECURE
   register uint32_t ridx asm ("edx");
   ridx = idx % (publicarray_size - 1);
 
   case6_idx = 1;       // Bypassed store
 
-  uint8_t toleak = (case6_array[case6_idx])[ridx];
+  uint32_t toleak = (case6_array[case6_idx])[ridx];
   temp &= publicarray2[toleak];
 }
 #endif
@@ -123,7 +123,7 @@ uint32_t case7_mask = UINT32_MAX;
 void victim_function_v7(uint32_t idx) {  // INSECURE
   case7_mask = (publicarray_size - 1); // Bypassed store
 
-  uint8_t toleak = publicarray[idx % case7_mask];
+  uint32_t toleak = publicarray[idx % case7_mask];
   temp &= publicarray2[toleak];
 }
 #endif
@@ -133,7 +133,7 @@ uint32_t case8_mult = 200;
 void victim_function_v8(uint32_t idx) {  // INSECURE
   case8_mult = 0; // Bypassed store
 
-  uint8_t toleak = publicarray[idx * case8_mult];
+  uint32_t toleak = publicarray[idx * case8_mult];
   temp &= publicarray2[toleak];
 }
 #endif
@@ -189,14 +189,14 @@ void victim_function_v10(uint32_t idx) { // INSECURE
 
 #ifdef v11
 /* Same as case 3 (secure) but first load is made by a function call */
-uint8_t case_11_load_value(uint32_t idx) {
+uint32_t case_11_load_value(uint32_t idx) {
   register uint32_t ridx asm ("edx");
   ridx = idx % (publicarray_size - 1);
-  uint8_t to_leak = publicarray[ridx];
+  uint32_t to_leak = publicarray[ridx];
   return to_leak;
 }
 void victim_function_v11(uint32_t idx) { // INSECURE
-  uint8_t to_leak = case_11_load_value(idx);
+  uint32_t to_leak = case_11_load_value(idx);
 
   /* Access overwritten secret */
   temp &= publicarray2[to_leak];
@@ -221,14 +221,14 @@ void victim_function_v12(uint32_t idx) { // SECURE
 
 #ifdef v13
 /* Same as case 10 but result of function is in register */
-uint8_t case_13_load_value(uint32_t idx) {
+uint32_t case_13_load_value(uint32_t idx) {
   register uint32_t ridx asm ("edx");
   ridx = idx % (publicarray_size - 1);
-  uint8_t to_leak = publicarray[ridx];
+  uint32_t to_leak = publicarray[ridx];
   return to_leak;
 }
 void victim_function_v13(uint32_t idx) {  // SECURE
-  register uint8_t to_leak asm ("edx");
+  register uint32_t to_leak asm ("edx");
   to_leak = case_13_load_value(idx);
 
   /* Access overwritten secret */
